plugin/kubernetes: add watch option
Disable watch by default (takes up mem) and allows user to override by
setting watch.

Update the README.md to match.

Fixes: #2156
